+++ drivers/gpu/nvgpu/common/pmu/clk/clk_vf_point.c
@@ -163,7 +163,7 @@ int clk_vf_point_pmu_setup(struct gk20a *g)
 
 static int clk_vf_point_construct_super(struct gk20a *g,
 					struct boardobj **ppboardobj,
-					u16 size, void *pargs)
+					size_t size, void *pargs)
 {
 	struct clk_vf_point *pclkvfpoint;
 	struct clk_vf_point *ptmpvfpoint =
@@ -246,7 +246,7 @@ static int _clk_vf_point_pmudatainit_freq(struct gk20a *g,
 
 static int clk_vf_point_construct_volt(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct clk_vf_point_volt *pclkvfpoint;
@@ -277,7 +277,7 @@ static int clk_vf_point_construct_volt(struct gk20a *g,
 
 static int clk_vf_point_construct_freq(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct clk_vf_point_freq *pclkvfpoint;
@@ -308,7 +308,7 @@ static int clk_vf_point_construct_freq(struct gk20a *g,
 
 static int clk_vf_point_construct_volt_35(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct clk_vf_point_volt *pclkvfpoint;
@@ -339,7 +339,7 @@ static int clk_vf_point_construct_volt_35(struct gk20a *g,
 
 static int clk_vf_point_construct_freq_35(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct clk_vf_point_freq *pclkvfpoint;
@@ -377,22 +377,22 @@ struct clk_vf_point *construct_clk_vf_point(struct gk20a *g, void *pargs)
 	switch (BOARDOBJ_GET_TYPE(pargs)) {
 	case CTRL_CLK_CLK_VF_POINT_TYPE_FREQ:
 		status = clk_vf_point_construct_freq(g, &board_obj_ptr,
-			(u16)sizeof(struct clk_vf_point_freq), pargs);
+			sizeof(struct clk_vf_point_freq), pargs);
 		break;
 
 	case CTRL_CLK_CLK_VF_POINT_TYPE_VOLT:
 		status = clk_vf_point_construct_volt(g, &board_obj_ptr,
-			(u16)sizeof(struct clk_vf_point_volt), pargs);
+			sizeof(struct clk_vf_point_volt), pargs);
 		break;
 
 	case CTRL_CLK_CLK_VF_POINT_TYPE_35_FREQ:
 		status = clk_vf_point_construct_freq_35(g, &board_obj_ptr,
-			(u16)sizeof(struct clk_vf_point_freq), pargs);
+			sizeof(struct clk_vf_point_freq), pargs);
 		break;
 
 	case CTRL_CLK_CLK_VF_POINT_TYPE_35_VOLT:
 		status = clk_vf_point_construct_volt_35(g, &board_obj_ptr,
-			(u16)sizeof(struct clk_vf_point_volt), pargs);
+			sizeof(struct clk_vf_point_volt), pargs);
 		break;
 
 	default:
